backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.10.17
httpx==0.27.2
beautifulsoup4==4.12.3
openai==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
numpy==1.26.4

